Move getAllOngoingChallengeMaps from challenges store to userchallenge store
not sure why i put in challenge in first place
