+++ frontend/src/store/challenges/actions.ts
@@ -1,17 +1,11 @@
-import {
-  ChallengeData,
-  ChallengeListData,
-  ChallengeMapData,
-} from '../../types/challenges';
+import { ChallengeData, ChallengeListData } from '../../types/challenges';
 import {
   REMOVE_CHALLENGE,
   RemoveChallengeAction,
   SAVE_CHALLENGE,
   SAVE_CHALLENGE_LIST,
-  SAVE_CHALLENGE_MAPS,
   SaveChallengeAction,
   SaveChallengeListAction,
-  SaveChallengeMapsAction,
 } from './types';
 
 export function saveChallengeList(
@@ -34,14 +28,5 @@ export function removeChallenge(challengeId: number): RemoveChallengeAction {
   return {
     type: REMOVE_CHALLENGE,
     challengeId,
-  };
-}
-
-export function saveChallengeMaps(
-  challengeMaps: ChallengeMapData[]
-): SaveChallengeMapsAction {
-  return {
-    type: SAVE_CHALLENGE_MAPS,
-    challengeMaps,
   };
 }
