+++ frontend/src/store/challenges/reducer.ts
@@ -5,7 +5,6 @@ import {
   REMOVE_CHALLENGE,
   SAVE_CHALLENGE,
   SAVE_CHALLENGE_LIST,
-  SAVE_CHALLENGE_MAPS,
 } from './types';
 
 const initialState: ChallengesState = {
@@ -27,10 +26,6 @@ const challengesReducer = produce(
       }
       case REMOVE_CHALLENGE: {
         delete draft.challenges[action.challengeId];
-        break;
-      }
-      case SAVE_CHALLENGE_MAPS: {
-        draft.challengeMaps = action.challengeMaps;
         break;
       }
     }
