+++ frontend/src/store/challenges/selectors.ts
@@ -1,10 +1,6 @@
 import { RootState } from '../index';
 import { ChallengesState } from './types';
-import {
-  ChallengeData,
-  ChallengeListData,
-  ChallengeMapData,
-} from '../../types/challenges';
+import { ChallengeData, ChallengeListData } from '../../types/challenges';
 
 function getLocalState(state: RootState): ChallengesState {
   return state.challenges;
@@ -19,8 +15,4 @@ export function getChallenge(
   challengeId: number
 ): ChallengeData {
   return getLocalState(state).challenges[challengeId];
-}
-
-export function getChallengeMaps(state: RootState): ChallengeMapData[] {
-  return getLocalState(state).challengeMaps;
 }
