+++ frontend/src/store/userchallenges/actions.ts
@@ -4,5 +4,6 @@ export const {
   updateAllUserChallengesData,
   updateOngoingUserChallengesListData,
   updateCompletedUserChallengesListData,
+  updateOngoingChallengeMapsData,
   resetUserChallenges,
 } = userChallengesSlice.actions;
