+++ frontend/src/store/userchallenges/operations.ts
@@ -1,5 +1,6 @@
 import { AnyAction } from 'redux';
 import { ThunkDispatch } from 'redux-thunk';
+import { ChallengeMapData } from 'types/challenges';
 import {
   CompletedUserChallengeListData,
   UserChallengeListData,
@@ -10,6 +11,7 @@ import { RootState } from '../index';
 import {
   updateAllUserChallengesData,
   updateCompletedUserChallengesListData,
+  updateOngoingChallengeMapsData,
   updateOngoingUserChallengesListData,
 } from './actions';
 
@@ -62,6 +64,14 @@ export function loadAllCompletedUserChallenges(
         data: [...userChallenges],
       })
     );
+  };
+}
+
+export function loadAllOngoingChallengeMaps(): OperationResult {
+  return async (dispatch: ThunkDispatch<RootState, undefined, AnyAction>) => {
+    const response = await api.map.getAllOngoingChallengeMaps();
+    const challengeMaps: ChallengeMapData[] = response.payload.data;
+    dispatch(updateOngoingChallengeMapsData({ data: [...challengeMaps] }));
   };
 }
